+++ rzauctionmonitor/parser/monitoring_tools/sqlwriter/P4InsertToSqlServer.cpp
@@ -37,8 +37,8 @@ struct DB_InsertItem {
 		uint32_t awaken_option_value[5];
 		int32_t awaken_option_data[5];
 		uint32_t random_type[10];
-		int64_t random_value_1[10];
-		int64_t random_value_2[10];
+		float random_value_1[10];
+		float random_value_2[10];
 		int32_t remain_time;
 		uint8_t elemental_effect_type;
 		int32_t elemental_effect_remain_time;
